rename to .js
for James :)
